+++ Nebula/src/Nebula/Renderer/Texture.h
@@ -35,7 +35,7 @@ namespace Nebula {
 
 		virtual uint32_t GetWidth() const = 0;
 		virtual uint32_t GetHeight() const = 0;
-		virtual uint32_t GetRendererID() const = 0;
+		virtual uint64_t GetRendererID() const = 0;
 		
 		virtual void SetData(Buffer data) = 0;
 		virtual void SetFilterNearest(bool nearest = true) = 0;
